requirements.txt
anyio==4.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
et_xmlfile==2.0.0
exceptiongroup==1.2.2
googlemaps==4.10.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
loguru==0.7.3
numpy==2.2.4
openpyxl==3.1.5
openrouteservice==2.3.3
pandas==2.2.3
pyfiglet==1.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
requests==2.32.3
six==1.17.0
sniffio==1.3.1
typing_extensions==4.13.0
tzdata==2025.2
urllib3==2.3.0
